Move utils to separate namespace
